split cluster/update into cluster/update_{master,worker}
