+++ pkg/cluster/update_worker.go
@@ -1,0 +1,107 @@
+package cluster
+
+import (
+	"context"
+
+	"github.com/Azure/azure-sdk-for-go/services/compute/mgmt/2018-06-01/compute"
+	"github.com/Azure/go-autorest/autorest/to"
+
+	"github.com/openshift/openshift-azure/pkg/api"
+	"github.com/openshift/openshift-azure/pkg/cluster/kubeclient"
+	"github.com/openshift/openshift-azure/pkg/config"
+)
+
+// updateWorkerAgentPool creates new VMs and removes old VMs one by one.
+func (u *simpleUpgrader) updateWorkerAgentPool(ctx context.Context, cs *api.OpenShiftManagedCluster, app *api.AgentPoolProfile) *api.PluginError {
+	ssName := config.GetScalesetName(app.Name)
+	ssHash, err := u.hasher.HashScaleSet(cs, app)
+	if err != nil {
+		return &api.PluginError{Err: err, Step: api.PluginStepUpdateWorkerAgentPoolHashScaleSet}
+	}
+
+	// store a list of all the VM instances now, so that if we end up creating
+	// new ones (in the crash recovery case, we might not), we can detect which
+	// they are
+	oldVMs, err := u.listVMs(ctx, cs.Properties.AzProfile.ResourceGroup, ssName)
+	if err != nil {
+		return &api.PluginError{Err: err, Step: api.PluginStepUpdateWorkerAgentPoolListVMs}
+	}
+
+	blob, err := u.readUpdateBlob()
+	if err != nil {
+		return &api.PluginError{Err: err, Step: api.PluginStepUpdateWorkerAgentPoolReadBlob}
+	}
+
+	// Filter out VMs that do not need to get upgraded. Should speed
+	// up retrying failed upgrades.
+	oldVMs = u.filterOldVMs(oldVMs, blob, ssHash)
+	vmsBefore := map[string]struct{}{}
+	for _, vm := range oldVMs {
+		vmsBefore[*vm.InstanceID] = struct{}{}
+	}
+
+	for _, vm := range oldVMs {
+		u.log.Infof("setting %s capacity to %d", ssName, app.Count+1)
+		future, err := u.ssc.Update(ctx, cs.Properties.AzProfile.ResourceGroup, ssName, compute.VirtualMachineScaleSetUpdate{
+			Sku: &compute.Sku{
+				Capacity: to.Int64Ptr(app.Count + 1),
+			},
+		})
+		if err != nil {
+			return &api.PluginError{Err: err, Step: api.PluginStepUpdateWorkerAgentPoolWaitForReady}
+		}
+
+		if err := future.WaitForCompletionRef(ctx, u.ssc.Client()); err != nil {
+			return &api.PluginError{Err: err, Step: api.PluginStepUpdateWorkerAgentPoolWaitForReady}
+		}
+
+		updatedList, err := u.listVMs(ctx, cs.Properties.AzProfile.ResourceGroup, ssName)
+		if err != nil {
+			return &api.PluginError{Err: err, Step: api.PluginStepUpdateWorkerAgentPoolListVMs}
+		}
+
+		// wait for newly created VMs to reach readiness (n.b. one alternative to
+		// this approach would be for the CSE to not return until the node is
+		// ready, but that is also problematic)
+		for _, updated := range updatedList {
+			if _, found := vmsBefore[*updated.InstanceID]; !found {
+				computerName := kubeclient.ComputerName(*updated.VirtualMachineScaleSetVMProperties.OsProfile.ComputerName)
+				u.log.Infof("waiting for %s to be ready", computerName)
+				err = u.kubeclient.WaitForReadyWorker(ctx, computerName)
+				if err != nil {
+					return &api.PluginError{Err: err, Step: api.PluginStepUpdateWorkerAgentPoolWaitForReady}
+				}
+				vmsBefore[*updated.InstanceID] = struct{}{}
+				blob.InstanceHashes[instanceName(*updated.Name)] = ssHash
+				if err := u.writeUpdateBlob(blob); err != nil {
+					return &api.PluginError{Err: err, Step: api.PluginStepUpdateWorkerAgentPoolUpdateBlob}
+				}
+			}
+		}
+
+		if err := u.deleteWorker(ctx, cs, app, *vm.InstanceID, kubeclient.ComputerName(*vm.VirtualMachineScaleSetVMProperties.OsProfile.ComputerName)); err != nil {
+			return &api.PluginError{Err: err, Step: api.PluginStepUpdateWorkerAgentPoolDeleteVMs}
+		}
+		delete(blob.InstanceHashes, instanceName(*vm.Name))
+		if err := u.writeUpdateBlob(blob); err != nil {
+			return &api.PluginError{Err: err, Step: api.PluginStepUpdateWorkerAgentPoolUpdateBlob}
+		}
+	}
+
+	return nil
+}
+
+func (u *simpleUpgrader) deleteWorker(ctx context.Context, cs *api.OpenShiftManagedCluster, app *api.AgentPoolProfile, instanceID string, nodeName kubeclient.ComputerName) error {
+	u.log.Infof("draining %s", nodeName)
+	if err := u.kubeclient.DrainAndDeleteWorker(ctx, nodeName); err != nil {
+		return err
+	}
+
+	u.log.Infof("deleting %s", nodeName)
+	future, err := u.vmc.Delete(ctx, cs.Properties.AzProfile.ResourceGroup, config.GetScalesetName(app.Name), instanceID)
+	if err != nil {
+		return err
+	}
+
+	return future.WaitForCompletionRef(ctx, u.vmc.Client())
+}
